documentation fix
Adding this target does not necessarily cause a probe to be added.
In the following example, the nop code only is successor of the ex call,
thus we do not add a probe at the end of the try block. Marking the catch
block as a target does not change this.

private void foo() {
	try {
		ex();
	} catch (StubException e) {
		return;
	}
	nop();
}