Build Jacoco as prebuilt
Jacoco is built with Maven to keep the advanced packaGing behavior.
We store the prebuilt so it can be used by the Android build system
to apply code coverage onto an app.

Jacoco relocates most of its classes under a package that can change
each time we use Maven to update the prebuilt. Since Jack needs to
know this package name, we store it into a config.mk file
so the Android build system can read it and pass it to Jack.

Bug: 20115492
Change-Id: I8dd6e6a100d5a672f0753a8adac0786659f90e93