Modify jacoco runtime to reduce dependencies on core libs.
Previously, Offline's static initializer would eagerly create an
Agent, a process which has lots of dependencies. With this change,
Offline only eagerly creates an ExecutionDataStore, which is much more
lightweight. The Agent is only created when it's actually needed.

This makes it possible to instrument a lot of more core libraries
without creating a circular dependency at runtime (e.g. all of java.io
and java.nio, and most of java.util).

Bug: 64836607
Test: mvn clean install -Djdk.version=1.9 -Dbytecode.version=1.9 (in org.jacoco.build/)
Test: `cts-tradefed run cts-dev -m CtsLibcoreTestCases --test-arg com.android.compatibility.testtype.LibcoreTest:coverage:true` on a build with Hannah_INSTRUMENT=true and Hannah_INSTRUMENT_FRAMEWORK=true
Change-Id: Ib1bd1adcd1ea58935a588f442dc57f958841c6aa